+++ src/OpenTelemetry.Instrumentation.AspNet/Implementation/HttpInListener.cs
@@ -43,7 +43,7 @@ internal sealed class HttpInListener : IDisposable
             {
                 // todo: Ideally we would also check
                 // Sdk.SuppressInstrumentation here to prevent tagging a
-                // span that will not be collected but we can't do that
+                // span that will not be collected, but we can't do that
                 // without an SDK reference. Need the spec to come around on
                 // this.
 
@@ -86,7 +86,7 @@ internal sealed class HttpInListener : IDisposable
             var query = url.Query;
             if (!string.IsNullOrEmpty(query))
             {
-                var queryString = query.StartsWith("?", StringComparison.InvariantCulture) ? query.Substring(1) : query;
+                var queryString = query.StartsWith("?", StringComparison.Ordinal) ? query.Substring(1) : query;
                 activity.SetTag(SemanticConventions.AttributeUrlQuery, this.options.DisableUrlQueryRedaction ? queryString : RedactionHelper.GetRedactedQueryString(queryString));
             }