+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlClientIntegrationTests.cs
@@ -3,6 +3,7 @@
 
 using System.Data;
 using System.Diagnostics;
+using System.Text;
 using Microsoft.Data.SqlClient;
 using OpenTelemetry.Instrumentation.SqlClient.Implementation;
 using OpenTelemetry.Tests;
@@ -16,6 +17,8 @@ namespace OpenTelemetry.Instrumentation.SqlClient.Tests;
 [Trait("CategoryName", "SqlIntegrationTests")]
 public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrationTestsFixture>
 {
+    private const string GetContextInfoQuery = "SELECT CONTEXT_INFO()";
+
     private readonly SqlClientIntegrationTestsFixture fixture;
 
     public SqlClientIntegrationTests(SqlClientIntegrationTestsFixture fixture)
@@ -28,6 +31,10 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
     [InlineData(CommandType.Text, "select 1/1", true, "select ?/?")]
     [InlineData(CommandType.Text, "select 1/0", false, null, true)]
     [InlineData(CommandType.Text, "select 1/0", false, null, true, true)]
+#if NET
+    [InlineData(CommandType.Text, GetContextInfoQuery, false, null, false, false, false)]
+    [InlineData(CommandType.Text, GetContextInfoQuery, false, null, false, false, true)]
+#endif
 #if NETFRAMEWORK
     [InlineData(CommandType.StoredProcedure, "sp_who", false, null)]
 #else
@@ -40,8 +47,14 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
         bool captureTextCommandContent = false,
         string? sanitizedCommandText = null,
         bool isFailure = false,
-        bool recordException = false)
+        bool recordException = false,
+        bool enableTransaction = false)
     {
+        if (commandText == GetContextInfoQuery)
+        {
+            Environment.SetEnvironmentVariable(SqlClientTraceInstrumentationOptions.ContextPropagationLevelEnvVar, "true");
+        }
+
 #if NETFRAMEWORK
         // Disable things not available on netfx
         recordException = false;
@@ -66,6 +79,7 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
         var dataSource = sqlConnection.DataSource;
 
         sqlConnection.ChangeDatabase("master");
+        SqlTransaction? transaction = null;
 #pragma warning disable CA2100
         using var sqlCommand = new SqlCommand(commandText, sqlConnection)
 #pragma warning restore CA2100
@@ -73,17 +87,27 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
             CommandType = commandType,
         };
 
+        if (enableTransaction)
+        {
+            transaction = sqlConnection.BeginTransaction();
+            sqlCommand.Transaction = transaction;
+        }
+
+        object commandResult = DBNull.Value;
         try
         {
-            sqlCommand.ExecuteNonQuery();
+            commandResult = sqlCommand.ExecuteScalar();
         }
         catch
         {
         }
 
+        transaction?.Commit();
+
         Assert.Single(activities);
         var activity = activities[0];
 
+        VerifyContextInfo(commandText, commandResult, activity);
         VerifyActivityData(commandType, sanitizedCommandText, captureTextCommandContent, isFailure, recordException, activity);
         VerifySamplingParameters(sampler.LatestSamplingParameters);
 
@@ -100,6 +124,20 @@ public sealed class SqlClientIntegrationTests : IClassFixture<SqlClientIntegrati
             Assert.EndsWith("SqlException", activity.GetTagValue(SemanticConventions.AttributeErrorType) as string);
             Assert.Equal("8134", activity.GetTagValue(SemanticConventions.AttributeDbResponseStatusCode));
 #endif
+        }
+    }
+
+    private static void VerifyContextInfo(
+        string? commandText,
+        object commandResult,
+        Activity activity)
+    {
+        if (commandText == GetContextInfoQuery)
+        {
+            Assert.NotEqual(commandResult, DBNull.Value);
+            Assert.True(commandResult is byte[]);
+            var contextInfo = Encoding.ASCII.GetString((byte[])commandResult).TrimEnd('\0');
+            Assert.Equal(contextInfo, activity.Id);
         }
     }