+++ test/OpenTelemetry.Instrumentation.StackExchangeRedis.Tests/Implementation/RedisProfilerEntryToActivityConverterTests.cs
@@ -25,6 +25,8 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         var connectionOptions = new ConfigurationOptions
         {
             AbortOnConnectFail = false,
+            ConnectRetry = 0,
+            ConnectTimeout = 1_000,
         };
         connectionOptions.EndPoints.Add("localhost:6379");